chore: address cleanup review feedback
Agent-Logs-Url: https://github.com/KTheMan/kennyandmorgan.com/sessions/297872c2-ccf0-49ee-a998-c1b79f35a247

+++ supabase/schema.sql
@@ -155,7 +155,7 @@ begin
     values (
         encode(digest(raw_token, 'sha256'), 'hex'),
         matched_level,
-        timezone('utc', now()) + make_interval(secs => greatest(session_ttl_ms, 60000) / 1000)
+        timezone('utc', now()) + make_interval(secs => (greatest(session_ttl_ms, 60000)::numeric / 1000))
     );
 
     return jsonb_build_object(
